feat(homePage): compatible mobile
