--- docs/components/components/Footer/config.ts
@@ -1,48 +1,0 @@
-export const footerGroups = [
-  {
-    title: 'Ant Design Mini',
-    items: [
-      {
-        title: 'GitHub',
-        link: 'https://github.com/ant-design/ant-design-mini',
-      },
-      {
-        title: '发布日志',
-        link: 'https://github.com/ant-design/ant-design-mini/releases',
-      },
-      {
-        title: '参与贡献',
-        link: 'https://github.com/ant-design/ant-design-mini/blob/main/.github/contributing.md',
-      },
-      {
-        title: '提问与反馈',
-        link: 'https://github.com/ant-design/ant-design-mini/issues/new/choose',
-      },
-    ],
-  },
-  {
-    title: 'Ant Design',
-    items: [
-      { title: 'Ant Design - 桌面端组件库', link: 'https://ant.design' },
-      {
-        title: 'Ant Design Mobile - 移动端组件库',
-        link: 'https://mobile.ant.design',
-      },
-      {
-        title: 'Ant Design Pro - 中台解决方案',
-        link: 'https://pro.ant.design',
-      },
-      { title: 'Ant Motion - 设计动效', link: 'https://motion.ant.design' },
-      { title: 'Ant Design Charts', link: 'https://charts.ant.design/' },
-    ],
-  },
-  {
-    title: '更多产品',
-    items: [
-      { title: 'Dumi - 组件/文档研发工具', link: 'https://d.umijs.org' },
-      { title: 'Kitchen - Sketch 工具集', link: 'https://kitchen.alipay.com' },
-      { title: 'AntV - 蚂蚁数据可视化方案', link: 'https://antv.vision' },
-      { title: 'Egg - 企业级 Node Web 开发框架', link: 'https://eggjs.org' },
-    ],
-  },
-]
