+++ docs/components/components/Footer/footerMobile.less
@@ -1,0 +1,20 @@
+@media screen and (max-width: 450px) {
+  .footerContainer {
+    .footerBody {
+      text-align: center;
+
+      .column {
+        width: 100vw;
+        margin-bottom: 72px;
+      }
+
+      .column:last-child {
+        margin-bottom: 0;
+      }
+    }
+
+    .copyright {
+      max-width: 100%;
+    }
+  }
+}
