+++ docs/components/components/Footer/index.local.less
@@ -1,42 +1,65 @@
-.footer {
-  margin-top: 60px;
-  background: #06080a;
+@import "./footerMobile.less";
 
-  .columns {
-    margin: 0 auto;
-    width: 1140px;
+.footerContainer {
+  background-color: #000;
+  padding-top: 60px;
+  color: #fff;
+
+  .footerBody {
     display: flex;
+    flex-wrap: wrap;
+    max-width: 1184px;
+    margin: 0 auto;
     justify-content: space-between;
-    padding: 80px 0 32px;
 
-    ul {
-      flex: 1;
-      padding: 0 36px 0 0;
+    .column {
+      .columnTile {
+        position: relative;
+        margin: 0 auto 24px;
+        font-weight: 500;
+        font-size: 16px;
+        color: #fff;
+      }
 
-      li {
-        font-size: 14px;
-        line-height: 22px;
-        margin-bottom: 12px;
-        list-style: none;
+      .columnItem {
+        margin: 12px 0;
 
-        a {
-          color: rgba(255, 255, 255, 0.9);
+        .columnItemCustomIcon {
+          width: 16px;
+          height: 16px;
+        }
+
+        .columnItemIcon {
+          margin-right: 6px;
+        }
+
+        .columnItemUrl {
+          color: #fff;
+          opacity: 0.9;
+          text-decoration: none;
+        }
+
+        .columnItemDescription {
+          color: #666;
         }
       }
-
-      li:first-child {
-        font-size: 16px;
-        color: #ffffff;
-        margin-bottom: 24px;
-      }
     }
+
   }
 
   .copyright {
-    padding: 12px 0;
-    font-size: 14px;
+    padding: 16px 0;
+    max-width: calc(100% - 120px);
+    margin: 60px auto 0;
+    font-size: 16px;
     color: rgba(255, 255, 255, 0.65);
     text-align: center;
-    line-height: 22px;
+    line-height: 32px;
+    border-top: 1px solid rgba(255, 255, 255, .25);
+
+    .copyrightLink {
+      color: #fff;
+      margin-left: 8px;
+    }
   }
 }
