+++ docs/components/components/Main/MainSection/index.local.less
@@ -60,8 +60,15 @@
 
       .publishDayAnimation {
         position: relative;
-        left: -14px;
+        left: -52px;
+        top: -18px;
       }
+    }
+
+    .resultImage {
+      position: absolute;
+      right: 80px;
+      top: 80px;
     }
   }
 }
@@ -76,7 +83,7 @@
 
     .imageContainer {
       position: absolute;
-      width: calc(100%-24px);
+      width: calc(100% - 24px);
       top: 140px;
       overflow: hidden;
 
@@ -88,12 +95,24 @@
         width: 128px;
         height: 112px;
         left: 18%;
-        top: 176px
+        top: 176px;
+
+        .publishDayAnimation {
+          top: -10px;
+          left: -28px;
+        }
+      }
+
+      .resultImage {
+        right: 50px;
+        top: 120px;
       }
     }
 
     .mainSectionButtonAction {
+      position: relative;
       margin-top: 290px;
+      z-index: 99;
     }
   }
 }
