+++ docs/components/components/Main/index.local.less
@@ -291,7 +291,6 @@
     }
   }
 
-
   @media screen and (max-width: 450px) {
     .mainSection {
       padding-top: 12px;
@@ -367,7 +366,6 @@
           width: 100% !important;
         }
       }
-
     }
   }
