+++ docs/components/globalMobile.less
@@ -1,0 +1,49 @@
+:global {
+
+  // 移动端
+  @media screen and (max-width: 450px) {
+
+    // 影响移动端所有页面
+    #root .__dumi-default-navbar {
+      padding: 0 58px;
+    }
+
+    .__dumi-default-search-input {
+      display: none;
+    }
+
+    // 影响移动端首页
+    .__dumi-default-layout[data-route="/"].__dumi-default-layout[data-gapless='true'] {
+      .__dumi-default-navbar {
+        position: sticky;
+        top: 0;
+        background-color: #1677ff;
+        opacity: 0.95;
+        box-shadow: 0px 1px 1px 0px #eeeeee40 !important;
+      }
+
+      .__dumi-default-navbar-toggle {
+        width: 16px;
+        height: 16px;
+        background: url(https://gw.alipayobjects.com/zos/bmw-prod/2a780b32-f3dd-4e13-8c27-2369d898bd76.svg) no-repeat center / cover
+      }
+
+      .__dumi-default-menu {
+        background-color: #1677ff;
+        opacity: 0.95;
+
+        .__dumi-default-menu-inner .__dumi-default-menu-mobile-area {
+          border-bottom: none;
+        }
+
+        a {
+          color: #fff !important;
+        }
+      }
+
+      .__dumi-default-menu[data-mode='site']::after {
+        transform: scaleX(0.2);
+      }
+    }
+  }
+}
