+++ docs/components/home.local.less
@@ -1,127 +1,11 @@
+@import "./globalPC.less";
+@import "./globalMobile.less";
+
 .homePage {
   overflow: hidden;
   background: #fff;
 
   a {
     text-decoration: none;
-  }
-}
-
-:global {
-
-  // 下面样式只影响首页
-  .__dumi-default-layout[data-route="/"].__dumi-default-layout[data-gapless='true'] {
-    padding-top: 0 !important;
-
-    .__dumi-default-navbar {
-      position: absolute;
-      max-width: 1152px;
-      margin: 0 auto;
-      background-color: transparent;
-      padding: 0 !important;
-    }
-
-    .__dumi-default-navbar-logo:not([data-plaintext]) {
-      // width: 233px;
-      height: 32px;
-      background-image: url(https://gw.alipayobjects.com/zos/bmw-prod/7c5fec71-34b2-4996-8384-89555714927c.svg) !important;
-    }
-
-    .__dumi-default-navbar nav>span>a {
-      color: #fff;
-    }
-
-    .__dumi-default-navbar nav>span>a:hover {
-      font-family: PingFangSC-Medium;
-      line-height: 22px;
-      color: #fff !important;
-    }
-
-    .__dumi-default-navbar nav>span:hover:before {
-      content: '';
-      position: absolute;
-      top: 0;
-      left: -2px;
-      right: -2px;
-      height: 2px;
-      background-color: #fff;
-      border-radius: 1px;
-    }
-
-    .__dumi-default-search-input {
-      border-radius: 8px !important;
-      color: #fff !important;
-      width: 200px;
-      height: 32px;
-      background: url(https://gw.alipayobjects.com/zos/bmw-prod/f4f3a6af-273c-4250-8cf3-f36e58448063.svg) no-repeat;
-      background-size: 100% 100%;
-    }
-
-    .__dumi-default-search-input::placeholder {
-      color: #ffffff80;
-    }
-
-    .__dumi-default-search-input:focus {
-      box-shadow: none;
-    }
-  }
-
-  @media screen and (max-width: 450px) {
-
-    // 影响移动端所有页面
-    #root .__dumi-default-navbar {
-      padding: 0 58px;
-    }
-
-    // 影响移动端首页
-    .__dumi-default-layout[data-route="/"].__dumi-default-layout[data-gapless='true'] {
-      .__dumi-default-navbar {
-        position: sticky;
-        top: 0;
-        background-color: #1677ff;
-        opacity: 0.95;
-        box-shadow: 0px 1px 1px 0px #eeeeee40 !important;
-      }
-
-      .__dumi-default-navbar-toggle {
-        width: 16px;
-        height: 16px;
-        background: url(https://gw.alipayobjects.com/zos/bmw-prod/2a780b32-f3dd-4e13-8c27-2369d898bd76.svg) no-repeat center / contain
-      }
-
-      .__dumi-default-menu {
-        background-color: #1677ff;
-        opacity: 0.95;
-
-        .__dumi-default-menu-inner .__dumi-default-menu-mobile-area {
-          border-bottom: none;
-        }
-
-        a {
-          color: #fff !important;
-        }
-      }
-
-      .__dumi-default-menu[data-mode='site']::after {
-        transform: scaleX(0.2);
-      }
-    }
-  }
-
-  // 下面影响所有页面
-  .__dumi-default-navbar nav>span>a svg {
-    display: none;
-  }
-
-  .__dumi-default-search-input {
-    vertical-align: middle !important;
-  }
-
-  .__dumi-default-search-input {
-    display: none;
-  }
-
-  .__dumi-default-navbar-logo {
-    background-position-x: center !important;
   }
 }
